Move ArtistsList to TS
